requirements.txt
faiss-gpu
numpy
torch
transformers
ollama
newspaper3k
lxml_html_clean